+++ test/projectile-legacy-test.el
@@ -190,10 +190,6 @@
   (should (string-prefix-p "git" (projectile-get-sub-projects-command 'git)))
   (should (string-empty-p (projectile-get-sub-projects-command 'none))))
 
-(ert-deftest projectile-test-files-via-ext-command ()
-  (should (not (projectile-files-via-ext-command "" "")))
-  (should (not (projectile-files-via-ext-command "" nil))))
-
 (ert-deftest projectile-test-setup-hook-functions-projectile-mode ()
   (noflet ((projectile--cleanup-known-projects () nil)
            (projectile-discover-projects-in-search-path () nil))